Error in Italian translation
There is a grammatical inconsistency between the two therms: in Italian we accord the adjectives as male or female so "Attivo" in italian is attributed to male type, "Non attiva" is attributed to female type. This commit is a fix for this inconsistency translating all in male type (correct for this kind of reference). All the Italian people will thank you so much when they notice the merge of this commit. Thank you